# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Document Assistant",
    "description": "Backend API for AI-powered document processing, semantic search, document management, and chat.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/documents/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get user documents",
        "description": "Return all documents uploaded by the authenticated user.",
        "operationId": "get_documents_api_documents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Documents Api Documents  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete document",
        "description": "Delete an authenticated user's document, its stored vectors, and its physical file.",
        "operationId": "delete_document_api_documents__document_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Document Api Documents  Document Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history/": {
      "get": {
        "tags": [
          "Chat History"
        ],
        "summary": "Get chat history",
        "description": "Return chat history belonging to the authenticated user. Results may optionally be restricted to one document.",
        "operationId": "get_history_api_history__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of chat records to return.",
              "default": 5,
              "title": "Limit"
            },
            "description": "Maximum number of chat records to return."
          },
          {
            "name": "document_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional document ID used to restrict history to one document.",
              "title": "Document Id"
            },
            "description": "Optional document ID used to restrict history to one document."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chat History"
        ],
        "summary": "Clear chat history",
        "description": "Delete chat history belonging to the authenticated user. Deletion may optionally be restricted to one document.",
        "operationId": "clear_history_api_history__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional document ID used to clear history for one document only.",
              "title": "Document Id"
            },
            "description": "Optional document ID used to clear history for one document only."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearChatHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatHistoryListResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryResponse"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "ChatHistoryListResponse",
        "description": "Collection of stored chat-history records."
      },
      "ChatHistoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "title": "Answer"
          },
          "mode": {
            "type": "string",
            "enum": [
              "document",
              "groq",
              "smart"
            ],
            "title": "Mode"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "question",
          "answer",
          "mode",
          "created_at",
          "updated_at"
        ],
        "title": "ChatHistoryResponse",
        "description": "Public representation of one stored chat interaction."
      },
...
      "ClearChatHistoryResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "deleted_records": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Deleted Records"
          }
        },
        "type": "object",
        "required": [
          "message",
          "deleted_records"
        ],
        "title": "ClearChatHistoryResponse",
        "description": "Response returned after clearing chat history."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}